+++ src/WireMock.Net.Testcontainers/WireMockContainerBuilder.cs
@@ -63,6 +63,29 @@ public sealed class WireMockContainerBuilder : ContainerBuilder<WireMockContaine
     }
 
     /// <summary>
+    /// Sets a custom WireMock.Net image for which to create the container.
+    /// </summary>
+    /// <param name="image">The image name.</param>
+    /// <returns>A configured instance of <see cref="WireMockContainerBuilder"/></returns>
+    [PublicAPI]
+    public new WireMockContainerBuilder WithImage(string image)
+    {
+        return WithCustomImage(image);
+    }
+
+    /// <summary>
+    /// Sets a custom WireMock.Net image for which to create the container.
+    /// </summary>
+    /// <param name="image">The image name.</param>
+    /// <returns>A configured instance of <see cref="WireMockContainerBuilder"/></returns>
+    [PublicAPI]
+    public WireMockContainerBuilder WithCustomImage(string image)
+    {
+        _imageOS ??= TestcontainersUtils.GetImageOSAsync.Value.GetAwaiter().GetResult();
+        return base.WithImage(image);
+    }
+
+    /// <summary>
     /// Set the admin username and password for the container (basic authentication).
     /// </summary>
     /// <param name="username">The admin username.</param>
@@ -207,9 +230,9 @@ public sealed class WireMockContainerBuilder : ContainerBuilder<WireMockContaine
         // In case the _imageOS is not set, determine it from the Image FullName.
         if (_imageOS == null)
         {
-            if (builder.DockerResourceConfiguration.Image.FullName.IndexOf("wiremock.net", StringComparison.OrdinalIgnoreCase) < 0)
+            if (builder.DockerResourceConfiguration.Image.FullName.IndexOf("wiremock", StringComparison.OrdinalIgnoreCase) < 0)
             {
-                throw new InvalidOperationException();
+                throw new InvalidOperationException("It's only possible to use a wiremock docker image.");
             }
 
             _imageOS = builder.DockerResourceConfiguration.Image.FullName.IndexOf("windows", StringComparison.OrdinalIgnoreCase) >= 0 ? OSPlatform.Windows : OSPlatform.Linux;