+++ test/WireMock.Net.Tests/Testcontainers/TestcontainersTests.Grpc.cs
@@ -205,6 +205,8 @@ public partial class TestcontainersTests
 
         var result = await httpClient.PostAsync("/__admin/mappings", new StringContent(mappingsJson, Encoding.UTF8, WireMockConstants.ContentTypeJson));
         result.EnsureSuccessStatusCode();
+
+        await Task.Delay(1000);
     }
 
     private static async Task<HelloReply> When_GrpcClient_Calls_SayHelloAsync(WireMockContainer wireMockContainer)